Update zip script/Fix errors.

+++ zip.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+#variables
+FILE=*.zip
+
 #get all files
 cp arch/arm/boot/zImage ../builds/AIK-Linux/split_img
 cp drivers/scsi/scsi_wait_scan.ko ../builds/system/lib/modules
@@ -14,15 +17,40 @@ cp image-new.img ..
 
 #Add boot.img to zip
 cd ..
-mv image-new.img boot.img
-cp *.zip previousbuilds/
-mv *.zip test.zip
-zip -g test.zip boot.img
-zip -g test.zip system/lib/modules/scsi_wait_scan.ko
-zip -g test.zip system/lib/modules/dhd.ko
+if [ -f $FILE ];
+then
+    echo "File $FILE exists."
+    mv image-new.img boot.img
+    echo "Backup to previousbuilds/"
+	cp *.zip previousbuilds/
+	mv *.zip test.zip
+	echo "Adding files to new zip..."
+	zip -g test.zip boot.img
+	zip -g test.zip system/lib/modules/scsi_wait_scan.ko
+	zip -g test.zip system/lib/modules/dhd.ko
 
-#Name new kernel.zip
-echo "Enter name of new zip: "
-read input_variable
-mv test.zip $input_variable.zip
-echo "$input_variable.zip is ready"
+	#Name new kernel.zip
+	echo "Enter name of new zip: "
+	read input_variable
+	mv test.zip $input_variable.zip
+	echo "$input_variable.zip is ready"
+else
+    echo "File $FILE does not exist."
+    echo "Copying base.zip from base folder."
+    cp base/base.zip test.zip
+    mv image-new.img boot.img
+	cp *.zip previousbuilds/
+	mv *.zip test.zip
+	echo "Adding files to new zip."
+	zip -g test.zip boot.img
+	zip -g test.zip system/lib/modules/scsi_wait_scan.ko
+	zip -g test.zip system/lib/modules/dhd.ko
+
+	#Name new kernel.zip
+	echo "Enter name of new zip: "
+	read input_variable
+	mv test.zip $input_variable.zip
+	echo "$input_variable.zip is ready"
+fi
+
+
